+++ docs/build-info.ts
@@ -1,1 +1,1 @@
-export const BUILD_TIME = "2026-05-16 09:58:56";
+export const BUILD_TIME = "2026-05-16 15:04:39";